Moved Credentials to Separate File; Removed Unused Import
